feat: store process key by process id and deployment key + enable lookup
Adds a new column family to store process definition keys by process id and deployment key.
The column family does not store the entire object, but just a foreign key to reduce the required storage space.

The column family will only be propagated for new deployments / "process created" events to ensure that existing events
are replayed deterministically. This is done by introducing a new event applier and record version.

Also, a new method is added to the process state that allows to find processes by process id and deployment key.
This will be needed to find the correct version of a called process if the new "deployment" binding type is used.
The method uses caches to reduce latency:
1. look up process definition key by process id and deployment key from a new cache introduced with this commit
2. look up deployed process by process definition key from the existing cache