feat: store decision key by decision id and deployment key + add lookup
Adds a new column family to store decision keys by decision id and deployment key.
The column family does not store the entire object, but just a foreign key to reduce the required storage space.

The column family will only be propagated for new deployments / "decision created" events to ensure that existing events
are replayed deterministically. This is done by introducing a new event applier and record version.

Also, a new method is added to the decision state that allows to find decisions by decision id and deployment key.
This will be needed to find the correct version of a called decision if the new "deployment" binding type is used.