feat: store form key by form id and deployment key + enable lookup
Adds a new column family to store form keys by form id and deployment key.
The column family does not store the entire object, but just a foreign key to reduce the required storage space.

The column family will only be propagated for new deployments / "form created" events to ensure that existing events are
replayed deterministically. This is done by introducing a new event applier and record version.

Also, a new method is added to the form state that allows to find forms by process id and deployment key.
This will be needed to find the correct version of a linked form if the new "deployment" binding type is used.